Revert "Worker Main 코드에 OOP 리팩토링 적용"
