+++ worker/src/main/scala/Main.scala
@@ -1,14 +1,13 @@
 import io.grpc.ServerBuilder
 import scala.concurrent.ExecutionContext
 import utils.{WorkerOptionUtils, PathUtils, SamplingUtils}
+import main.MasterManager
 import server.{WorkerServiceImpl}
 import global.WorkerState
 import worker.WorkerService.WorkerServiceGrpc
 import scala.concurrent.Future
 import common.utils.SystemUtils
 import global.ConnectionManager
-import main.RegisterManager
-import main.SampleManager
 
 object Main extends App {
   implicit val ec: ExecutionContext = ExecutionContext.global
@@ -44,11 +43,39 @@ object Main extends App {
 
   server.start()
 
+  val workerIp = SystemUtils.getLocalIp.getOrElse {
+    println("Failed to get local IP address")
+    sys.exit(1)
+  }
+  val ramMb = SystemUtils.getRamMb
+  val port = server.getPort
+
   ConnectionManager.initMasterChannel(masterIp, masterPort)
 
-  new RegisterManager().start(server.getPort)
+  val client = new MasterManager()
+  client.registerWorker(workerIp, port, ramMb)
 
-  new SampleManager().start()
+  // Start sampling in a separate thread
+  val samplingPhase = Future {
+    try {
+      val samples = SamplingUtils.sampleFromInputs(inputDirs).getOrElse {
+        println("Warning: Sampling failed")
+        sys.exit(1)
+      }
+      
+      val success = client.sampling(workerIp, samples)
+      
+      if (success) {
+        println("Samples sent successfully. Waiting for range assignment...")
+      } else {
+        println("Failed to send samples to master")
+      }
+    } catch {
+      case e: Exception =>
+        println(s"Error during sampling: ${e.getMessage}")
+        e.printStackTrace()
+    }
+  }
 
   ConnectionManager.shutdownAllChannels()
